requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.4.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.2
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
junitparser==1.4.0
MarkupSafe==1.1.1
more-itertools==7.2.0
openapi-spec-validator==0.2.8
packaging==19.2
pkg-resources==0.0.0
pluggy==0.13.0
psycopg2-binary==2.8.4
py==1.8.0
PyMySQL==0.9.3
pyparsing==2.4.3
pyrsistent==0.15.5
pytest==5.2.2
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.10
swagger-ui-bundle==0.0.6
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
WTForms==2.2.1
zipp==0.6.0
